Agegar qery obtener todas las clases por servicio

+++ src/main/java/com/bodyFitnessGym/repository/ClaseRepository.java
@@ -10,5 +10,5 @@ import com.bodyFitnessGym.model.entity.Clase;
 public interface ClaseRepository extends CrudRepository<Clase, Long> {
 
 	@Query(value = "SELECT * FROM CLASE c WHERE c.servicio_id = ?1", nativeQuery = true)
-	Collection<Clase> findAllClaseByIdServicio(Integer status);
+	Collection<Clase> findAllClaseByIdServicio(Long status);
 }
